split app.js into focused ES modules
extension/app.js was a single 2,532-line file that mixed Chrome API
wrappers, storage CRUD, formatting helpers, audio/animation effects,
HTML generators, the render pipeline, and the click handler. Pulling
each concern into its own module so the file is easier to navigate
and individual pieces are easier to find and reason about.

New modules in extension/:
  - format.js       pure string / domain / date helpers
  - effects.js      Web Audio swoosh, confetti, FLIP, toast
  - storage.js      chrome.storage.local CRUD + workspace migration
  - tabs.js         chrome.tabs/windows wrappers + private openTabs snapshot
  - render.js       HTML string builders + ICONS + buildLiveWorkspaceMap
  - icon-picker.js  workspace emoji picker popover + EMOJI_CATEGORIES

extension/app.js is now a 1,155-line orchestrator: render pipeline,
event delegation / click handler, live tab sync, init.

extension/index.html: <script src="app.js"> -> <script type="module">
to enable native ES module imports. No build step, no bundler.
MV3 page contexts already support modules so no manifest change.

Behavior is preserved end-to-end. The only non-mechanical tweak:
buildLiveWorkspaceMap now takes openTabs as a parameter instead of
closing over a module-level global, since render.js and tabs.js are
now separate. Both call sites in app.js pass getOpenTabs().

HANDOFF.txt at the repo root documents the new module layout, the
dependency graph, the verification checklist, and a "where things
live now" index.
