+++ extension/effects.js
@@ -1,0 +1,236 @@
+/* ================================================================
+   effects.js — audio, confetti, FLIP, toast, empty-state
+
+   All side-effect-only helpers that touch the DOM, Web Audio,
+   and requestAnimationFrame. No app state, no storage, no chrome.* APIs.
+   ================================================================ */
+
+'use strict';
+
+
+/**
+ * playCloseSound()
+ *
+ * Plays a clean "swoosh" sound when tabs are closed.
+ * Built entirely with the Web Audio API — no sound files needed.
+ * A filtered noise sweep that descends in pitch, like air moving.
+ */
+export function playCloseSound() {
+  try {
+    const ctx = new (window.AudioContext || window.webkitAudioContext)();
+    const t = ctx.currentTime;
+
+    // Swoosh: shaped white noise through a sweeping bandpass filter
+    const duration = 0.25;
+    const buffer = ctx.createBuffer(1, ctx.sampleRate * duration, ctx.sampleRate);
+    const data = buffer.getChannelData(0);
+
+    // Generate noise with a natural envelope (quick attack, smooth decay)
+    for (let i = 0; i < data.length; i++) {
+      const pos = i / data.length;
+      // Envelope: ramps up fast in first 10%, then fades out smoothly
+      const env = pos < 0.1 ? pos / 0.1 : Math.pow(1 - (pos - 0.1) / 0.9, 1.5);
+      data[i] = (Math.random() * 2 - 1) * env;
+    }
+
+    const source = ctx.createBufferSource();
+    source.buffer = buffer;
+
+    // Bandpass filter sweeps from high to low — creates the "swoosh" character
+    const filter = ctx.createBiquadFilter();
+    filter.type = 'bandpass';
+    filter.Q.value = 2.0;
+    filter.frequency.setValueAtTime(4000, t);
+    filter.frequency.exponentialRampToValueAtTime(400, t + duration);
+
+    // Volume
+    const gain = ctx.createGain();
+    gain.gain.setValueAtTime(0.15, t);
+    gain.gain.exponentialRampToValueAtTime(0.001, t + duration);
+
+    source.connect(filter).connect(gain).connect(ctx.destination);
+    source.start(t);
+
+    setTimeout(() => ctx.close(), 500);
+  } catch {
+    // Audio not supported — fail silently
+  }
+}
+
+/**
+ * shootConfetti(x, y)
+ *
+ * Shoots a burst of colorful confetti particles from the given screen
+ * coordinates (typically the center of a card being closed).
+ * Pure CSS + JS, no libraries.
+ */
+export function shootConfetti(x, y) {
+  const colors = [
+    '#c8713a', // amber
+    '#e8a070', // amber light
+    '#5a7a62', // sage
+    '#8aaa92', // sage light
+    '#5a6b7a', // slate
+    '#8a9baa', // slate light
+    '#d4b896', // warm paper
+    '#b35a5a', // rose
+  ];
+
+  const particleCount = 17;
+
+  for (let i = 0; i < particleCount; i++) {
+    const el = document.createElement('div');
+
+    const isCircle = Math.random() > 0.5;
+    const size = 5 + Math.random() * 6; // 5–11px
+    const color = colors[Math.floor(Math.random() * colors.length)];
+
+    el.style.cssText = `
+      position: fixed;
+      left: ${x}px;
+      top: ${y}px;
+      width: ${size}px;
+      height: ${size}px;
+      background: ${color};
+      border-radius: ${isCircle ? '50%' : '2px'};
+      pointer-events: none;
+      z-index: 9999;
+      transform: translate(-50%, -50%);
+      opacity: 1;
+    `;
+    document.body.appendChild(el);
+
+    // Physics: random angle and speed for the outward burst
+    const angle   = Math.random() * Math.PI * 2;
+    const speed   = 60 + Math.random() * 120;
+    const vx      = Math.cos(angle) * speed;
+    const vy      = Math.sin(angle) * speed - 80; // bias upward
+    const gravity = 200;
+
+    const startTime = performance.now();
+    const duration  = 700 + Math.random() * 200; // 700–900ms
+
+    function frame(now) {
+      const elapsed  = (now - startTime) / 1000;
+      const progress = elapsed / (duration / 1000);
+
+      if (progress >= 1) { el.remove(); return; }
+
+      const px = vx * elapsed;
+      const py = vy * elapsed + 0.5 * gravity * elapsed * elapsed;
+      const opacity = progress < 0.5 ? 1 : 1 - (progress - 0.5) * 2;
+      const rotate  = elapsed * 200 * (isCircle ? 0 : 1);
+
+      el.style.transform = `translate(calc(-50% + ${px}px), calc(-50% + ${py}px)) rotate(${rotate}deg)`;
+      el.style.opacity = opacity;
+
+      requestAnimationFrame(frame);
+    }
+
+    requestAnimationFrame(frame);
+  }
+}
+
+/**
+ * animateCardOut(card)
+ *
+ * Smoothly removes a mission card: fade + scale down, then confetti.
+ * After the animation, checks if the grid is now empty.
+ */
+export function animateCardOut(card) {
+  if (!card) return;
+
+  const rect = card.getBoundingClientRect();
+  shootConfetti(rect.left + rect.width / 2, rect.top + rect.height / 2);
+
+  card.classList.add('closing');
+  setTimeout(() => {
+    card.remove();
+    checkAndShowEmptyState();
+  }, 300);
+}
+
+/**
+ * showToast(message)
+ *
+ * Brief pop-up notification at the bottom of the screen.
+ */
+export function showToast(message) {
+  const toast = document.getElementById('toast');
+  document.getElementById('toastText').textContent = message;
+  toast.classList.add('visible');
+  setTimeout(() => toast.classList.remove('visible'), 2500);
+}
+
+/**
+ * checkAndShowEmptyState()
+ *
+ * Shows a cheerful "Inbox zero" message when all domain cards are gone.
+ */
+export function checkAndShowEmptyState() {
+  const missionsEl = document.getElementById('openTabsMissions');
+  if (!missionsEl) return;
+
+  const remaining = missionsEl.querySelectorAll('.mission-card:not(.closing)').length;
+  if (remaining > 0) return;
+
+  missionsEl.innerHTML = `
+    <div class="missions-empty-state">
+      <div class="empty-checkmark">
+        <svg xmlns="http://www.w3.org/2000/svg" fill="none" viewBox="0 0 24 24" stroke-width="1.5" stroke="currentColor">
+          <path stroke-linecap="round" stroke-linejoin="round" d="m4.5 12.75 6 6 9-13.5" />
+        </svg>
+      </div>
+      <div class="empty-title">Inbox zero, but for tabs.</div>
+      <div class="empty-subtitle">You're free.</div>
+    </div>
+  `;
+
+  const countEl = document.getElementById('openTabsSectionCount');
+  if (countEl) countEl.textContent = '0 domains';
+}
+
+
+/**
+ * flipDomainCards(container)
+ *
+ * FLIP animation helper: captures old positions of every mission card,
+ * returns a `playFlip()` closure. Caller updates innerHTML between
+ * the capture and the play, and the closure animates each card from
+ * its old position to its new one.
+ */
+export function flipDomainCards(container) {
+  const oldRects = new Map();
+  container.querySelectorAll('.mission-card[data-domain]').forEach(card => {
+    oldRects.set(card.dataset.domain, card.getBoundingClientRect());
+  });
+
+  return function playFlip() {
+    container.querySelectorAll('.mission-card[data-domain]').forEach(card => {
+      const domain = card.dataset.domain;
+      const newRect = card.getBoundingClientRect();
+      const oldRect = oldRects.get(domain);
+
+      if (oldRect) {
+        const dx = oldRect.left - newRect.left;
+        const dy = oldRect.top  - newRect.top;
+        if (dx === 0 && dy === 0) return;
+
+        card.style.transition = 'none';
+        card.style.transform  = `translate(${dx}px, ${dy}px)`;
+
+        requestAnimationFrame(() => requestAnimationFrame(() => {
+          card.style.transition = '';
+          card.style.transform  = '';
+        }));
+      } else {
+        card.style.transition = 'none';
+        card.style.opacity    = '0';
+        requestAnimationFrame(() => requestAnimationFrame(() => {
+          card.style.transition = 'opacity 0.3s ease';
+          card.style.opacity    = '';
+        }));
+      }
+    });
+  };
+}
