+++ extension/format.js
@@ -1,0 +1,229 @@
+/* ================================================================
+   format.js — pure string / domain / date helpers
+
+   No DOM, no chrome.* APIs, no module state. Safe to import anywhere.
+   ================================================================ */
+
+'use strict';
+
+
+/* ----------------------------------------------------------------
+   DOMAIN & TITLE CLEANUP HELPERS
+   ---------------------------------------------------------------- */
+
+// Map of known hostnames → friendly display names.
+export const FRIENDLY_DOMAINS = {
+  'github.com':           'GitHub',
+  'www.github.com':       'GitHub',
+  'gist.github.com':      'GitHub Gist',
+  'youtube.com':          'YouTube',
+  'www.youtube.com':      'YouTube',
+  'music.youtube.com':    'YouTube Music',
+  'x.com':                'X',
+  'www.x.com':            'X',
+  'twitter.com':          'X',
+  'www.twitter.com':      'X',
+  'reddit.com':           'Reddit',
+  'www.reddit.com':       'Reddit',
+  'old.reddit.com':       'Reddit',
+  'substack.com':         'Substack',
+  'www.substack.com':     'Substack',
+  'medium.com':           'Medium',
+  'www.medium.com':       'Medium',
+  'linkedin.com':         'LinkedIn',
+  'www.linkedin.com':     'LinkedIn',
+  'stackoverflow.com':    'Stack Overflow',
+  'www.stackoverflow.com':'Stack Overflow',
+  'news.ycombinator.com': 'Hacker News',
+  'google.com':           'Google',
+  'www.google.com':       'Google',
+  'mail.google.com':      'Gmail',
+  'docs.google.com':      'Google Docs',
+  'drive.google.com':     'Google Drive',
+  'calendar.google.com':  'Google Calendar',
+  'meet.google.com':      'Google Meet',
+  'gemini.google.com':    'Gemini',
+  'chatgpt.com':          'ChatGPT',
+  'www.chatgpt.com':      'ChatGPT',
+  'chat.openai.com':      'ChatGPT',
+  'claude.ai':            'Claude',
+  'www.claude.ai':        'Claude',
+  'code.claude.com':      'Claude Code',
+  'notion.so':            'Notion',
+  'www.notion.so':        'Notion',
+  'figma.com':            'Figma',
+  'www.figma.com':        'Figma',
+  'slack.com':            'Slack',
+  'app.slack.com':        'Slack',
+  'discord.com':          'Discord',
+  'www.discord.com':      'Discord',
+  'wikipedia.org':        'Wikipedia',
+  'en.wikipedia.org':     'Wikipedia',
+  'amazon.com':           'Amazon',
+  'www.amazon.com':       'Amazon',
+  'netflix.com':          'Netflix',
+  'www.netflix.com':      'Netflix',
+  'spotify.com':          'Spotify',
+  'open.spotify.com':     'Spotify',
+  'vercel.com':           'Vercel',
+  'www.vercel.com':       'Vercel',
+  'npmjs.com':            'npm',
+  'www.npmjs.com':        'npm',
+  'developer.mozilla.org':'MDN',
+  'arxiv.org':            'arXiv',
+  'www.arxiv.org':        'arXiv',
+  'huggingface.co':       'Hugging Face',
+  'www.huggingface.co':   'Hugging Face',
+  'producthunt.com':      'Product Hunt',
+  'www.producthunt.com':  'Product Hunt',
+  'xiaohongshu.com':      'RedNote',
+  'www.xiaohongshu.com':  'RedNote',
+  'local-files':          'Local Files',
+};
+
+export function friendlyDomain(hostname) {
+  if (!hostname) return '';
+  if (FRIENDLY_DOMAINS[hostname]) return FRIENDLY_DOMAINS[hostname];
+
+  if (hostname.endsWith('.substack.com') && hostname !== 'substack.com') {
+    return capitalize(hostname.replace('.substack.com', '')) + "'s Substack";
+  }
+  if (hostname.endsWith('.github.io')) {
+    return capitalize(hostname.replace('.github.io', '')) + ' (GitHub Pages)';
+  }
+
+  let clean = hostname
+    .replace(/^www\./, '')
+    .replace(/\.(com|org|net|io|co|ai|dev|app|so|me|xyz|info|us|uk|co\.uk|co\.jp)$/, '');
+
+  return clean.split('.').map(part => capitalize(part)).join(' ');
+}
+
+export function capitalize(str) {
+  if (!str) return '';
+  return str.charAt(0).toUpperCase() + str.slice(1);
+}
+
+export function stripTitleNoise(title) {
+  if (!title) return '';
+  // Strip leading notification count: "(2) Title" — that's a browser/page
+  // unread badge, not part of the actual page name.
+  title = title.replace(/^\(\d+\+?\)\s*/, '');
+  // Clean X/Twitter format
+  title = title.replace(/\s+on X:\s*/, ': ');
+  title = title.replace(/\s*\/\s*X\s*$/, '');
+  return title.trim();
+}
+
+export function cleanTitle(title, hostname) {
+  if (!title || !hostname) return title || '';
+
+  const friendly = friendlyDomain(hostname);
+  const domain   = hostname.replace(/^www\./, '');
+  const seps     = [' - ', ' | ', ' — ', ' · ', ' – '];
+
+  for (const sep of seps) {
+    const idx = title.lastIndexOf(sep);
+    if (idx === -1) continue;
+    const suffix     = title.slice(idx + sep.length).trim();
+    const suffixLow  = suffix.toLowerCase();
+    if (
+      suffixLow === domain.toLowerCase() ||
+      suffixLow === friendly.toLowerCase() ||
+      suffixLow === domain.replace(/\.\w+$/, '').toLowerCase() ||
+      domain.toLowerCase().includes(suffixLow) ||
+      friendly.toLowerCase().includes(suffixLow)
+    ) {
+      const cleaned = title.slice(0, idx).trim();
+      if (cleaned.length >= 5) return cleaned;
+    }
+  }
+  return title;
+}
+
+export function smartTitle(title, url) {
+  if (!url) return title || '';
+  let pathname = '', hostname = '';
+  try { const u = new URL(url); pathname = u.pathname; hostname = u.hostname; }
+  catch { return title || ''; }
+
+  const titleIsUrl = !title || title === url || title.startsWith(hostname) || title.startsWith('http');
+
+  if ((hostname === 'x.com' || hostname === 'twitter.com' || hostname === 'www.x.com') && pathname.includes('/status/')) {
+    const username = pathname.split('/')[1];
+    if (username) return titleIsUrl ? `Post by @${username}` : title;
+  }
+
+  if (hostname === 'github.com' || hostname === 'www.github.com') {
+    const parts = pathname.split('/').filter(Boolean);
+    if (parts.length >= 2) {
+      const [owner, repo, ...rest] = parts;
+      if (rest[0] === 'issues' && rest[1]) return `${owner}/${repo} Issue #${rest[1]}`;
+      if (rest[0] === 'pull'   && rest[1]) return `${owner}/${repo} PR #${rest[1]}`;
+      if (rest[0] === 'blob' || rest[0] === 'tree') return `${owner}/${repo} — ${rest.slice(2).join('/')}`;
+      if (titleIsUrl) return `${owner}/${repo}`;
+    }
+  }
+
+  if ((hostname === 'www.youtube.com' || hostname === 'youtube.com') && pathname === '/watch') {
+    if (titleIsUrl) return 'YouTube Video';
+  }
+
+  if ((hostname === 'www.reddit.com' || hostname === 'reddit.com' || hostname === 'old.reddit.com') && pathname.includes('/comments/')) {
+    const parts  = pathname.split('/').filter(Boolean);
+    const subIdx = parts.indexOf('r');
+    if (subIdx !== -1 && parts[subIdx + 1]) {
+      if (titleIsUrl) return `r/${parts[subIdx + 1]} post`;
+    }
+  }
+
+  return title || url;
+}
+
+
+/* ----------------------------------------------------------------
+   DATE / GREETING HELPERS
+   ---------------------------------------------------------------- */
+
+/**
+ * timeAgo(dateStr)
+ *
+ * Converts an ISO date string into a human-friendly relative time.
+ * "2026-04-04T10:00:00Z" → "2 hrs ago" or "yesterday"
+ */
+export function timeAgo(dateStr) {
+  if (!dateStr) return '';
+  const then = new Date(dateStr);
+  const now  = new Date();
+  const diffMins  = Math.floor((now - then) / 60000);
+  const diffHours = Math.floor((now - then) / 3600000);
+  const diffDays  = Math.floor((now - then) / 86400000);
+
+  if (diffMins < 1)   return 'just now';
+  if (diffMins < 60)  return diffMins + ' min ago';
+  if (diffHours < 24) return diffHours + ' hr' + (diffHours !== 1 ? 's' : '') + ' ago';
+  if (diffDays === 1) return 'yesterday';
+  return diffDays + ' days ago';
+}
+
+/**
+ * getGreeting() — "Good morning / afternoon / evening"
+ */
+export function getGreeting() {
+  const hour = new Date().getHours();
+  if (hour < 12) return 'Good morning';
+  if (hour < 17) return 'Good afternoon';
+  return 'Good evening';
+}
+
+/**
+ * getDateDisplay() — "Friday, April 4, 2026"
+ */
+export function getDateDisplay() {
+  return new Date().toLocaleDateString('en-US', {
+    weekday: 'long',
+    year:    'numeric',
+    month:   'long',
+    day:     'numeric',
+  });
+}
