+++ extension/tabs.js
@@ -1,0 +1,198 @@
+/* ================================================================
+   tabs.js — chrome.tabs / chrome.windows wrappers
+
+   Wraps every direct Chrome tab API call. The current `openTabs`
+   snapshot is private to this module; callers read it via
+   getOpenTabs(). Every mutation function refreshes the snapshot
+   so callers always see post-mutation state.
+   ================================================================ */
+
+'use strict';
+
+
+// All open tabs — populated by fetchOpenTabs(). Private to this module.
+let openTabs = [];
+
+/**
+ * getOpenTabs() — current snapshot of all browser tabs.
+ *
+ * Returns the same reference each time fetchOpenTabs() runs, so
+ * callers should NOT mutate it. Read-only.
+ */
+export function getOpenTabs() {
+  return openTabs;
+}
+
+/**
+ * fetchOpenTabs()
+ *
+ * Reads all currently open browser tabs directly from Chrome.
+ * Sets the isTabOut flag so we can identify Tab Out's own pages.
+ */
+export async function fetchOpenTabs() {
+  try {
+    const extensionId = chrome.runtime.id;
+    // The new URL for this page is now index.html (not newtab.html)
+    const newtabUrl = `chrome-extension://${extensionId}/index.html`;
+
+    const tabs = await chrome.tabs.query({});
+    openTabs = tabs.map(t => ({
+      id:            t.id,
+      url:           t.url,
+      title:         t.title,
+      windowId:      t.windowId,
+      active:        t.active,
+      index:         t.index,   // position in the tab strip — used for workspace ordering & tab-restore
+      // Opera GX exposes these; other Chromium browsers leave them undefined.
+      // No icon field is exposed — workspace icons come from chrome.storage.local
+      // and are managed via the in-UI emoji picker (the pencil on each pill).
+      workspaceId:   t.workspaceId,
+      workspaceName: t.workspaceName,
+      // Flag Tab Out's own pages so we can detect duplicate new tabs
+      isTabOut:      t.url === newtabUrl || t.url === 'chrome://newtab/',
+    }));
+  } catch {
+    // chrome.tabs API unavailable (shouldn't happen in an extension page)
+    openTabs = [];
+  }
+}
+
+/**
+ * closeTabsByUrls(urls)
+ *
+ * Closes all open tabs whose hostname matches any of the given URLs.
+ * After closing, re-fetches the tab list to keep our state accurate.
+ *
+ * Special case: file:// URLs are matched exactly (they have no hostname).
+ */
+export async function closeTabsByUrls(urls) {
+  if (!urls || urls.length === 0) return;
+
+  // Separate file:// URLs (exact match) from regular URLs (hostname match)
+  const targetHostnames = [];
+  const exactUrls = new Set();
+
+  for (const u of urls) {
+    if (u.startsWith('file://')) {
+      exactUrls.add(u);
+    } else {
+      try { targetHostnames.push(new URL(u).hostname); }
+      catch { /* skip unparseable */ }
+    }
+  }
+
+  const allTabs = await chrome.tabs.query({});
+  const toClose = allTabs
+    .filter(tab => {
+      const tabUrl = tab.url || '';
+      if (tabUrl.startsWith('file://') && exactUrls.has(tabUrl)) return true;
+      try {
+        const tabHostname = new URL(tabUrl).hostname;
+        return tabHostname && targetHostnames.includes(tabHostname);
+      } catch { return false; }
+    })
+    .map(tab => tab.id);
+
+  if (toClose.length > 0) await chrome.tabs.remove(toClose);
+  await fetchOpenTabs();
+}
+
+/**
+ * closeTabsExact(urls)
+ *
+ * Closes tabs by exact URL match (not hostname). Used for landing pages
+ * so closing "Gmail inbox" doesn't also close individual email threads.
+ */
+export async function closeTabsExact(urls) {
+  if (!urls || urls.length === 0) return;
+  const urlSet = new Set(urls);
+  const allTabs = await chrome.tabs.query({});
+  const toClose = allTabs.filter(t => urlSet.has(t.url)).map(t => t.id);
+  if (toClose.length > 0) await chrome.tabs.remove(toClose);
+  await fetchOpenTabs();
+}
+
+/**
+ * focusTab(url)
+ *
+ * Switches Chrome to the tab with the given URL (exact match first,
+ * then hostname fallback). Also brings the window to the front.
+ */
+export async function focusTab(url) {
+  if (!url) return;
+  const allTabs = await chrome.tabs.query({});
+  const currentWindow = await chrome.windows.getCurrent();
+
+  // Try exact URL match first
+  let matches = allTabs.filter(t => t.url === url);
+
+  // Fall back to hostname match
+  if (matches.length === 0) {
+    try {
+      const targetHost = new URL(url).hostname;
+      matches = allTabs.filter(t => {
+        try { return new URL(t.url).hostname === targetHost; }
+        catch { return false; }
+      });
+    } catch {}
+  }
+
+  if (matches.length === 0) return;
+
+  // Prefer a match in a different window so it actually switches windows
+  const match = matches.find(t => t.windowId !== currentWindow.id) || matches[0];
+  await chrome.tabs.update(match.id, { active: true });
+  await chrome.windows.update(match.windowId, { focused: true });
+}
+
+/**
+ * closeDuplicateTabs(urls, keepOne)
+ *
+ * Closes duplicate tabs for the given list of URLs.
+ * keepOne=true → keep one copy of each, close the rest.
+ * keepOne=false → close all copies.
+ */
+export async function closeDuplicateTabs(urls, keepOne = true) {
+  const allTabs = await chrome.tabs.query({});
+  const toClose = [];
+
+  for (const url of urls) {
+    const matching = allTabs.filter(t => t.url === url);
+    if (keepOne) {
+      const keep = matching.find(t => t.active) || matching[0];
+      for (const tab of matching) {
+        if (tab.id !== keep.id) toClose.push(tab.id);
+      }
+    } else {
+      for (const tab of matching) toClose.push(tab.id);
+    }
+  }
+
+  if (toClose.length > 0) await chrome.tabs.remove(toClose);
+  await fetchOpenTabs();
+}
+
+
+/* ----------------------------------------------------------------
+   HELPER: filter out browser-internal pages
+   ---------------------------------------------------------------- */
+
+/**
+ * getRealTabs()
+ *
+ * Returns tabs that are real web pages — no chrome://, extension
+ * pages, about:blank, etc. Operates on the cached snapshot from
+ * the last fetchOpenTabs() call.
+ */
+export function getRealTabs() {
+  return openTabs.filter(t => {
+    const url = t.url || '';
+    return (
+      !url.startsWith('chrome://') &&
+      !url.startsWith('chrome-extension://') &&
+      !url.startsWith('about:') &&
+      !url.startsWith('edge://') &&
+      !url.startsWith('brave://')
+    );
+  });
+}
